Added new 'Hexa' class to support hex+alpha values
The new `Hexa` class initialisation can take either a standard hex color value (defaults to opacity of 1) or an eight-digit string.  Internally, the alpha value is stored as a float to make it more compatible with the other color classes, but when represented as a string it'll be converted back to a hax value.

Notable changes:

* New Hexa class added with tests
* Factory changed as it can't just assume something starting with `#` is the Hex class
* Other color class types updated to add `toHexa` method
* Test for other color types updated to test the `toHexa` methods